requirements.txt
openai
elevenlabslib
pydub
python-dotenv
git+https://github.com/openai/whisper.git